Move test to the correct file
